Major code refactor
Code was a monolith around TableView and the pick entry point with the table mode.
Now the code was properly refactored into 4 different major components, namely the
table, the selectors, the output processors and the view itself. The table is
composed of the Table class and the SelectableTable class, the first holds the data
structure behind the (duh) table itself and the latter holds logic for selected
cells and subtables. Selectors are responsible for the 2 different selection modes:
single-selection where the user selects one cell at a time, and multi-selection
where the user composes the selection by means of inserting and removing subtables
from the selection. The output processors are responsible for generating the output
that will go to stdout when the user hits enter, there are 2: the list and the
table, self-explanatory. Finally, the View class is responsible for managing the
"UI", user inputs, etc.